Replace two complex zeros with two real poles
Intead of pulling the root locus of the 250Hz poles to 100Hz with two
zeros on the j*omega axis, we add two poles on the negative real axis.
This forces the 250Hz root locus to the left half plane.

The new poles has a branch cut, passing the desired domainating pole
position (0.68 damping, ~100Hz oscillation freq). It is conditionally
stable with gain < 1e9, so please use it with caution.